doc:Brute变量修改位置
